@@ -10,13 +10,19 @@
         "watch-poll": "mix watch -- --watch-options-poll=1000"
     },
     "dependencies": {
+        "@inertiajs/inertia-vue": "^0.7.2",
         "bootstrap-vue": "^2.21.2",
+        "vue-router": "^3.5.3",
         "vuetify": "^2.5.10"
     },
     "devDependencies": {
+        "@inertiajs/inertia": "^0.10.0",
+        "@inertiajs/inertia-vue3": "^0.5.1",
+        "@inertiajs/progress": "^0.2.6",
         "@mdi/font": "^6.4.95",
-        "@tailwindcss/forms": "^0.3.1",
-        "@tailwindcss/typography": "^0.4.0",
+        "@tailwindcss/forms": "^0.2.1",
+        "@tailwindcss/typography": "^0.3.0",
+        "@vue/compiler-sfc": "^3.0.5",
         "alpinejs": "^3.0.6",
         "axios": "^0.21",
         "bootstrap": "^4.6.0",
@@ -28,15 +34,15 @@
         "lodash": "^4.17.19",
         "popper.js": "^1.16.1",
         "postcss": "^8.1.14",
-        "postcss-import": "^14.0.1",
+        "postcss-import": "^12.0.1",
         "resolve-url-loader": "^3.1.2",
         "sass": "^1.32.11",
         "sass-loader": "^11.0.1",
         "source-map-loader": "^3.0.0",
-        "tailwindcss": "^2.2.2",
-        "vue": "^2.6.12",
+        "tailwindcss": "^2.0.1",
+        "vue": "^3.0.5",
         "vue-cli-plugin-vuetify": "~2.4.3",
-        "vue-loader": "^15.9.8",
+        "vue-loader": "^16.1.2",
         "vue-template-compiler": "^2.6.12",
         "vuetify-loader": "^1.7.0"
     }
