@@ -15,7 +15,7 @@ return [
     |
     */
 
-    'stack' => 'livewire',
+    'stack' => 'inertia',
 
     /*
      |--------------------------------------------------------------------------
