@@ -5,8 +5,6 @@ namespace Tests\Feature;
 use App\Models\User;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Laravel\Jetstream\Features;
-use Laravel\Jetstream\Http\Livewire\ApiTokenManager;
-use Livewire\Livewire;
 use Tests\TestCase;
 
 class CreateApiTokenTest extends TestCase
@@ -25,15 +23,13 @@ class CreateApiTokenTest extends TestCase
             $this->actingAs($user = User::factory()->create());
         }
 
-        Livewire::test(ApiTokenManager::class)
-                    ->set(['createApiTokenForm' => [
-                        'name' => 'Test Token',
-                        'permissions' => [
-                            'read',
-                            'update',
-                        ],
-                    ]])
-                    ->call('createApiToken');
+        $response = $this->post('/user/api-tokens', [
+            'name' => 'Test Token',
+            'permissions' => [
+                'read',
+                'update',
+            ],
+        ]);
 
         $this->assertCount(1, $user->fresh()->tokens);
         $this->assertEquals('Test Token', $user->fresh()->tokens->first()->name);
